Add files via upload
This project is a 2" spectroscopy sensor consisting of a AS7421 and a AS7343. Testing for astronomical purposes. Measurement ranges are 750nm - 1050nm(64 channels) and 405nm - 855nm(14 channels). Some channels are duplicated and others will not be used; such as flicker. I2C address for each sensor is 0x64 and 0x39. An external light source will be used for calibration and completed device is intended for eyepiece projection with an intended exit pupil of 2.5mm.